+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v2.7.0
+- Added support for custom pelts
 - Added support for converting audio files to AudioClip objects
 - Added support for adding custom tracks to the Gramophone